Update GlobeView.swift
Now tests for map failure (eg, cannot be found) and then tries with the standard map. Only then, if the standard map cannot be found, will a fatal error occur. This is to help the app recover itself when a bad map is specified, the app crashes, then the user restarts it.