Update RegionEntryController.swift
Added MainDelegate to be able to talk to the main window. Main window is focused on start up (even though this window will remain topmost) to make sure things behave as the user expects. Assign a default region name. Added more handling for determining which click is for which coordinate.